lfm2: the conv output is bf16, and a bisect switch for bring-up
The out projection takes its activation as bf16, the way every GEMV here does
and the way dx.py's attention output already did. The core was writing f32 and
the sequence was streaming one 4096-byte element of it, so the projection read
half a vector and misread its dtype. y and the state leave on separate fifos
now, because they are different types.

Two declarations had to follow: the ExternalFunction arg types (the MLIR
verifier caught that one) and the fifo element type (nothing catches that one -
it just produces noise). Worth knowing that the conv output crosses three
declarations and only the middle one is checked.

SC_PASSTHROUGH makes the core emit C unchanged, so the expected value is
trivial and a bring-up run can tell a plumbing fault from an arithmetic one.

NOT WORKING YET. Against the fp64 oracle one conv layer gives corr 0.012, and
the passthrough build returns all-NaN logits while every input the core sees is
sane (|C| = 8.6, max 2.1, nothing near overflow). So the NaN is made on device
and the fault is in the fabric, not the math in sc.h - which compiles, and whose
arithmetic is pinned against the reference in test_short_conv_kernel.py.

OPEN-SHORT-CONV-KERNEL.

+++ open_kernels/designs/short_conv/sc_elem.cc
@@ -6,7 +6,7 @@ void short_conv_step(const float *__restrict Bv, const float *__restrict Cv,
                      const float *__restrict uv, const bfloat16 *__restrict w0,
                      const bfloat16 *__restrict w1, const bfloat16 *__restrict w2,
                      const float *__restrict s0, const float *__restrict s1,
-                     float *__restrict y, float *__restrict n0, float *__restrict n1) {
+                     bfloat16 *__restrict y, float *__restrict n0, float *__restrict n1) {
   short_conv_elem(Bv, Cv, uv, w0, w1, w2, s0, s1, y, n0, n1);
 }
 }
